Move all files into a single src dir
