+++ src/bin/server.js
@@ -7,8 +7,8 @@ import summary from 'server-summary';
 import httpNdjson from 'http-ndjson';
 import cliclopts from 'cliclopts';
 import minimist from 'minimist';
-import VERSION from '../lib/version.js';
-import collapsify from '../lib/node.js';
+import VERSION from '../version.js';
+import collapsify from '../node.js';
 
 const allowedArgs = [
   {
