+++ src/collapsers/binary.js
