+++ src/collapsers/javascript.js
