+++ src/collapsify.js
