+++ src/plugins/parse5-flatten-external-style.js
