+++ src/plugins/parse5-flatten-image.js
