+++ src/plugins/parse5-flatten-inline-style.js
