+++ src/plugins/parse5-flatten-script.js
