+++ src/utils/data-uri.js
