+++ src/utils/parse5-async-rewriter.js
